IOC builder:  * Fix Hy8515 default.  * CMS IOC limits should be floats.  * Add all ArgInfo decorated entities to a DecoratedCallables list.  * Tweak to mks937a
git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/iocbuilder@28676 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ iocbuilder/defaults/cmsIon.py
@@ -9,8 +9,8 @@ class cmsIon(Substitution, AutoProtocol):
     ArgInfo = [
         ('device', str, 'Device Prefix'),
         ('port', None, 'Asyn port'),
-        ('high', int, 'Integrated dose (since reset) HIGH value', 50),
-        ('hihi', int, 'Integrated dose (since reset) HIHI value', 100)        
+        ('high', float, 'Integrated dose (since reset) HIGH value', 50),
+        ('hihi', float, 'Integrated dose (since reset) HIHI value', 100)        
     ]    
     XMLObjects = ['port']    
     # Substitution attributes
